Add module and rename service object
