--- app/services/executable_service.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module ExecutableService
-  def execute(*args)
-    new(*args).execute
-  end
-end
