+++ app/services/service.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Service
+  def call(*args)
+    new(*args).call
+  end
+end
