+++ spec/factories/teams.rb
@@ -3,8 +3,8 @@
 FactoryBot.define do
   factory :team do
     association :created_by, factory: :user
-    name 'My team'
-    description 'Lorem ipsum dolor sit amet, consectetuer adipiscing eli.'
-    space_taken 1048576
+    sequence(:name) { |n| "My team-#{n}" }
+    description { Faker::Lorem.sentence }
+    space_taken { 1048576 }
   end
 end
