+++ spec/services/experiments/generate_workflow_image_service_spec.rb
@@ -2,18 +2,18 @@
 
 require 'rails_helper'
 
-describe WorkflowImageGenerator do
+describe Experiments::GenerateWorkflowImageService do
   let(:experiment) { create :experiment_with_tasks }
   let(:params) { { experiment_id: experiment.id } }
 
   context 'when succeed' do
     it 'succeed? returns true' do
-      expect(described_class.execute(params).succeed?).to be_truthy
+      expect(described_class.call(params).succeed?).to be_truthy
     end
 
     it 'worklfow image of experiment is updated' do
       old_filename = experiment.workflowimg_file_name
-      described_class.execute(params)
+      described_class.call(params)
       experiment.reload
 
       expect(experiment.workflowimg_file_name).not_to be == old_filename
